#926: Split up tournament helper into standard helper functions and layout engine
